+++ newSwiftUIApp/ScheduleViews.swift
@@ -1,0 +1,84 @@
+//
+//  ScheduleViews.swift
+//  newSwiftUIApp
+//
+//  Created by Ido Shoshani on 1/8/20.
+//  Copyright © 2020 Ido Shoshani. All rights reserved.
+//
+
+import SwiftUI
+
+struct ContentViewD: View {
+@ObservedObject var viewRouter: ViewRouter
+
+var body: some View {
+    
+    TabView {
+        
+        ScheduleView()
+        
+        .tabItem {
+            VStack {
+                Image(systemName: "clock")
+                Text("Schedule")
+                }
+        }.tag(1)
+        
+        HomeworkView()
+        
+        .tabItem {
+            VStack {
+                Image(systemName: "pencil")
+                Text("Homework")
+                }
+        }.tag(2)
+        
+        
+    }
+    
+    }
+}
+
+struct ScheduleView: View {
+
+var body: some View {
+        Text("Schedule View")
+    }
+}
+
+struct HomeworkView: View {
+
+var body: some View {
+        Text("Homework View")
+    }
+}
+
+struct ContentView_PreviewsD: PreviewProvider {
+    static var previews: some View {
+        ContentViewD(viewRouter: ViewRouter())
+    }
+}
+
+struct ScheduleView_Previews: PreviewProvider {
+    static var previews: some View {
+        ScheduleView()
+    }
+}
+
+struct HomeworkView_Previews: PreviewProvider {
+    static var previews: some View {
+        HomeworkView()
+    }
+}
+
+
+
+
+/*
+ VStack {
+       Text("All Done!").fontWeight(.heavy)
+     Button(action: {rwt.writeFile(writeString: "", fileName: "Save6"); self.viewRouter.currentPage = "page1"}) {
+         Text("Reset").fontWeight(.heavy).padding()
+     }
+ } .onAppear {classCount = 0}
+ */
